requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs
Babel==2.12.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.1
Django
django-cors-headers==3.13.0
django-filter
django-sequences==2.6
django-simple-history==3.1.1
djangorestframework==3.13.1
drf-yasg
execnet==1.9.0
gunicorn
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2
MarkupSafe==2.1.0
packaging==21.3
pandas==1.5.3
pluggy==1.0.0
psycopg2-binary
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.7
pytest
pytest-django
pytest-forked==1.4.0
pytest-xdist==2.5.0
pytz==2021.3
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.8
python-dateutil
num2words
xhtml2pdf==0.2.11
Numpy_financial
openpyxl==3.1.1
factory-boy==3.3.0
responses==0.23.1
django-extensions==3.2.3
pypdf==3.17.0
Pillow==10.1.0
PyMuPDF==1.23.7
django-enumfield==3.1
pytest-sugar
pytest-cov
drf-writable-nested==0.7.0
